change main

+++ src/Main.java
@@ -3,7 +3,7 @@ public class Main {
         BmiService service = new BmiService();
         double heightM = 1.87;
         int weightKG = 98;
-        int Bmi = service.calculate((int) weightKG, heightM);
+        int Bmi = service.calculate(weightKG, heightM);
         System.out.println(Bmi);
     }
 }
